updated todos

+++ src/wasm/math/types/color.h
@@ -28,11 +28,11 @@ namespace engine
 		float r, g, b, a;
 		
 		color() : r(0), g(0), b(0), a(1) {}
-		color(double v) : r(v), g(v), b(v), a(1) {}
+		color(float v) : r(v), g(v), b(v), a(1) {}
 		color(vec2 const& uv) : r(uv.x), g(uv.y), b(0), a(1) {}
 		color(vec3 const& rgb) : r(rgb.x), g(rgb.y), b(rgb.z), a(1) {}
 		color(vec4 const& rgba) : r(rgba.x), g(rgba.y), b(rgba.z), a(rgba.w) {}
-		color(double r, double g, double b, double a) : r(r), g(g), b(b), a(a) {}
+		color(float r, float g, float b, float a) : r(r), g(g), b(b), a(a) {}
 		
 		colorRGBint toRGBint() {
 			return {r * 255 + ROUNDUP_EPSILON, g * 255 + ROUNDUP_EPSILON, b * 255 + ROUNDUP_EPSILON};
@@ -68,6 +68,9 @@ namespace engine
 			return str.str();
 		}
 	};
+	
+	// @TODO:
+	// - convert between formats
 }
 
 #endif
